refactor(transcription): move the datasource factory into the package
Config -> engine mapping is the package's own concern:
TranscriptionConfig (the resolved config vocabulary) and
createTranscriptionDataSource move from the app data layer into
app_transcription, with the connection timeout injected by the caller
instead of reading an app constant. The app keeps only the theme-config
resolution in FeatureAccess; the factory tests move to the package and
the mapper tests to test/mappers.
