+++ lib/data/data.dart
@@ -21,4 +21,3 @@ export 'package_info.dart';
 export 'push_environment.dart';
 export 'secure_storage.dart';
 export 'session_cleanup_worker.dart';
-export 'transcription_datasource_factory.dart';
