+++ lib/models/feature_access/feature_access.dart
@@ -20,4 +20,4 @@ export 'sip_presence_config.dart';
 export 'supported_feature.dart';
 export 'system_notifications_config.dart';
 export 'terms_config.dart';
-export 'transcription_config.dart';
+export 'package:app_transcription/app_transcription.dart' show TranscriptionConfig, TranscriptionMode;
