+++ packages/data/app_transcription/lib/app_transcription.dart
@@ -1,6 +1,8 @@
 export 'package:_http_client/_http_client.dart' show TrustedCertificates;
 
 export 'src/local_whisper_transcription_datasource.dart';
+export 'src/transcription_config.dart';
+export 'src/transcription_datasource_factory.dart';
 export 'src/remote_whisper_transcription_datasource.dart';
 export 'src/transcription_datasource.dart';
 export 'src/transcription_service.dart';
