+++ packages/data/app_transcription/test/transcription_datasource_factory_test.dart
@@ -1,10 +1,8 @@
-import 'package:flutter_test/flutter_test.dart';
-
 import 'dart:typed_data';
 
-import 'package:webtrit_phone/data/data.dart';
-import 'package:webtrit_phone/models/models.dart';
-import 'package:webtrit_phone/theme/theme.dart';
+import 'package:flutter_test/flutter_test.dart';
+
+import 'package:app_transcription/app_transcription.dart';
 
 class _NoopTranscriptionStore implements TranscriptionStore {
   @override
@@ -124,37 +122,6 @@ void main() {
 
       expect(service.isEnabled, isTrue);
       expect(source.disposeCalls, 0);
-    });
-  });
-
-  group('TranscriptionMapper', () {
-    test('copies the transcription section from the app config', () {
-      const appConfig = AppConfig(
-        transcription: AppConfigTranscription(
-          mode: 'remote',
-          language: 'en',
-          local: AppConfigTranscriptionLocal(model: 'small'),
-          remote: AppConfigTranscriptionRemote(url: 'https://stt.example.com/v1', apiKey: 'key', model: 'large-v3'),
-        ),
-      );
-
-      final config = TranscriptionMapper.map(appConfig);
-
-      expect(config.mode, 'remote');
-      expect(config.language, 'en');
-      expect(config.localModel, 'small');
-      expect(config.remoteUrl, 'https://stt.example.com/v1');
-      expect(config.remoteApiKey, 'key');
-      expect(config.remoteModel, 'large-v3');
-    });
-
-    test('defaults to the disabled mode', () {
-      final config = TranscriptionMapper.map(const AppConfig());
-
-      expect(TranscriptionMode.fromName(config.mode), TranscriptionMode.disabled);
-      expect(config.language, isNull);
-      expect(config.localModel, 'base');
-      expect(config.remoteUrl, isNull);
     });
   });
 }
