+++ test/mappers/transcription_mapper_test.dart
@@ -1,0 +1,37 @@
+import 'package:flutter_test/flutter_test.dart';
+
+import 'package:webtrit_phone/data/data.dart';
+import 'package:webtrit_phone/theme/theme.dart';
+
+void main() {
+  group('TranscriptionMapper', () {
+    test('copies the transcription section from the app config', () {
+      const appConfig = AppConfig(
+        transcription: AppConfigTranscription(
+          mode: 'remote',
+          language: 'en',
+          local: AppConfigTranscriptionLocal(model: 'small'),
+          remote: AppConfigTranscriptionRemote(url: 'https://stt.example.com/v1', apiKey: 'key', model: 'large-v3'),
+        ),
+      );
+
+      final config = TranscriptionMapper.map(appConfig);
+
+      expect(config.mode, 'remote');
+      expect(config.language, 'en');
+      expect(config.localModel, 'small');
+      expect(config.remoteUrl, 'https://stt.example.com/v1');
+      expect(config.remoteApiKey, 'key');
+      expect(config.remoteModel, 'large-v3');
+    });
+
+    test('defaults to the disabled mode', () {
+      final config = TranscriptionMapper.map(const AppConfig());
+
+      expect(TranscriptionMode.fromName(config.mode), TranscriptionMode.disabled);
+      expect(config.language, isNull);
+      expect(config.localModel, 'base');
+      expect(config.remoteUrl, isNull);
+    });
+  });
+}
